convert sampleRandomNumbers test to vitest
